POSMA-288 #time 3h

+++ app/src/main/java/id/sisi/postoko/view/ui/addsales/AddSaleActivity.kt
@@ -107,7 +107,7 @@ class AddSaleActivity : BaseActivity() {
         listProduct.forEach { product ->
 //            logE("selectedTotal : ${product.isSelected}")
             if (product.isSelected){
-                total += (product.sale_qty * product.price)
+                total += (product.sale_qty * product.price.toInt())
             }
         }
         return total
